Controller Layer를 Scala 기반으로 전환한다. #221 scala.Long과 java.lang.Long 에서 발생하는 오류 수정함.
