+++ src/main/java/net/slipp/service/tag/TagService.java
@@ -18,6 +18,7 @@ import net.slipp.repository.tag.TagRepository;
 import net.slipp.repository.tag.TaggedHistoryRepository;
 import net.slipp.service.MailService;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -139,11 +140,25 @@ public class TagService {
 		}
     	return latestTags;
     }
-
-    public Tag saveTag(Tag tag) {
-        return tagRepository.save(tag);
-    }
     
+    public Tag saveTag(String name, Long parentTag) {
+    	if (StringUtils.isBlank(name)) {
+    		throw new IllegalArgumentException("태그명을 입력해 주세요.");
+    	}
+    	
+    	Tag originalTag = tagRepository.findByName(name);
+    	if (originalTag != null) {
+    		throw new IllegalArgumentException(name + " 태그는 이미 존재합니다.");
+    	}
+    	
+    	Tag parent = null;
+    	if (parentTag != null) {
+    		parent = tagRepository.findOne(parentTag);
+   	    }
+    	
+        return tagRepository.save(Tag.pooledTag(name, parent));
+    }
+
     public Tag findTagByName(String name) {
         return tagRepository.findByName(name);
     }
